SMBIOS: Automatically detect HHG Platform
HHG Platform was introduced with Lahaina later in the release cycle. It lacks a proper platform type, and instead makes use of the HDK type with specific subtype values. On Lahaina, these subtype values are 1 and 2. It is confirmed on official Kailua kernel sources that HHG is a dedicated platform. It also would not make sense to merge it with HDKs due to numerous differences. Detect HHG and override the type accordingly.