Build: Drastically decrease the required time to build everything
Having different **command line args** __severely__ impacts the amount of time required to build all targets which share **the same code half of the time** as it invalidates the previous build results and rebuilds **everything**.

Mitigate this severe issue (both time consuming and financially a disaster) by mitigating it, firstly, with the secureboot variants.

Building with SbOn first in CI, and then Off is a confirmed valid scenario
We do not change the cmd args between both runs!